Split painter module in smaller ones
