+++ gaphas/freehand.py
@@ -12,7 +12,8 @@ See: http://stevehanov.ca/blog/index.php?id=33 and
 from math import sqrt
 from random import Random
 
-from gaphas.painter import Context, Painter
+from gaphas.canvas import Context
+from gaphas.painter import Painter
 
 
 class FreeHandCairoContext:
